+++ index.ts
@@ -1,4 +1,0 @@
-export {GridsterModule} from './src/gridster.module';
-export {GridsterComponent} from './src/gridster.component';
-export {IGridsterOptions} from './src/IGridsterOptions';
-export {IGridsterDraggableOptions} from './src/IGridsterDraggableOptions';