+++ src/index.ts
@@ -1,0 +1,6 @@
+export {GridsterModule} from './gridster.module';
+export {GridsterComponent} from './gridster.component';
+export {GridsterItemComponent} from './gridster-item/gridster-item.component';
+export {GridsterItemPrototypeDirective} from './gridster-prototype/gridster-item-prototype.directive';
+export {IGridsterOptions} from './IGridsterOptions';
+export {IGridsterDraggableOptions} from './IGridsterDraggableOptions';